requirements.txt
Flask
pymongo[srv]
Flask-PyMongo
pytest
python-dotenv